pov-gen-agent/pyproject.toml
[project]
name = "pov-gen-agent"
version = "0.1.0"
description = "POV-based cold email generator using LangGraph and DeepSeek-V3"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2.0",
    "langchain>=0.3.0,<0.4.0",
    "langchain-openai>=0.2.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
